Reorganize package as murgtools with getdata and utils subpackages
- Rename getdatatestbed/ to murgtools/getdata/
- Move testbedutils/ to murgtools/utils/
- Create top-level murgtools/__init__.py with exports
- Update all import statements in source and test files
- Update setup.py, pyproject.toml, MANIFEST.in for new structure
- Update README.md and CLAUDE.md documentation

New usage:
  from murgtools.getdata import getObs
  from murgtools.utils import geoprocess
