+++ murgtools/getdata/getOutsideData.py
