+++ murgtools/utils/waveLib.py
