+++ tests/test_integration.py
@@ -8,8 +8,8 @@ from unittest.mock import MagicMock, patch, PropertyMock
 class TestGetnc:
     """Tests for the getnc function with mocked network."""
 
-    @patch('getdatatestbed.getDataFRF.socket.gethostbyname')
-    @patch('getdatatestbed.getDataFRF.nc.Dataset')
+    @patch('murgtools.getdata.getDataFRF.socket.gethostbyname')
+    @patch('murgtools.getdata.getDataFRF.nc.Dataset')
     def test_getnc_selects_chl_server_for_external_ip(self, mock_dataset, mock_gethostbyname):
         """Test that getnc selects CHL server for non-FRF IPs."""
         mock_gethostbyname.return_value = '192.168.1.1'
@@ -22,7 +22,7 @@ class TestGetnc:
         mock_ds.__getitem__ = MagicMock(return_value=mock_time)
         mock_dataset.return_value = mock_ds
 
-        from getdatatestbed.getDataFRF import getnc
+        from murgtools.getdata.getDataFRF import getnc
 
         result = getnc('oceanography/waves/waverider-26m.ncml', 'getObs')
 
@@ -30,8 +30,8 @@ class TestGetnc:
         call_args = mock_dataset.call_args[0][0]
         assert 'chldata.erdc.dren.mil' in call_args or 'chlthredds' in call_args
 
-    @patch('getdatatestbed.getDataFRF.socket.gethostbyname')
-    @patch('getdatatestbed.getDataFRF.nc.Dataset')
+    @patch('murgtools.getdata.getDataFRF.socket.gethostbyname')
+    @patch('murgtools.getdata.getDataFRF.nc.Dataset')
     def test_getnc_selects_frf_server_for_internal_ip(self, mock_dataset, mock_gethostbyname):
         """Test that getnc selects FRF server for internal IPs."""
         mock_gethostbyname.return_value = '134.164.129.50'
@@ -43,7 +43,7 @@ class TestGetnc:
         mock_ds.__getitem__ = MagicMock(return_value=mock_time)
         mock_dataset.return_value = mock_ds
 
-        from getdatatestbed.getDataFRF import getnc
+        from murgtools.getdata.getDataFRF import getnc
 
         result = getnc('oceanography/waves/waverider-26m.ncml', 'getObs', server='FRF')
 
@@ -51,8 +51,8 @@ class TestGetnc:
         call_args = mock_dataset.call_args[0][0]
         assert '134.164.129.55' in call_args
 
-    @patch('getdatatestbed.getDataFRF.socket.gethostbyname')
-    @patch('getdatatestbed.getDataFRF.nc.Dataset')
+    @patch('murgtools.getdata.getDataFRF.socket.gethostbyname')
+    @patch('murgtools.getdata.getDataFRF.nc.Dataset')
     def test_getnc_drills_to_monthly_file(self, mock_dataset, mock_gethostbyname):
         """Test that getnc drills down to monthly file when dates are in same month."""
         mock_gethostbyname.return_value = '192.168.1.1'
@@ -64,7 +64,7 @@ class TestGetnc:
         mock_ds.__getitem__ = MagicMock(return_value=mock_time)
         mock_dataset.return_value = mock_ds
 
-        from getdatatestbed.getDataFRF import getnc
+        from murgtools.getdata.getDataFRF import getnc
 
         d1 = DT.datetime(2020, 1, 5)
         d2 = DT.datetime(2020, 1, 15)
@@ -82,8 +82,8 @@ class TestGetnc:
         assert '2020' in call_args
         assert '202001' in call_args or '01.nc' in call_args
 
-    @patch('getdatatestbed.getDataFRF.socket.gethostbyname')
-    @patch('getdatatestbed.getDataFRF.nc.Dataset')
+    @patch('murgtools.getdata.getDataFRF.socket.gethostbyname')
+    @patch('murgtools.getdata.getDataFRF.nc.Dataset')
     def test_getnc_retries_on_io_error(self, mock_dataset, mock_gethostbyname):
         """Test that getnc retries on IOError."""
         mock_gethostbyname.return_value = '192.168.1.1'
@@ -97,9 +97,9 @@ class TestGetnc:
 
         mock_dataset.side_effect = [IOError("Network error"), IOError("Network error"), mock_ds]
 
-        from getdatatestbed.getDataFRF import getnc
+        from murgtools.getdata.getDataFRF import getnc
 
-        with patch('getdatatestbed.getDataFRF.time.sleep'):  # Don't actually sleep in tests
+        with patch('murgtools.getdata.getDataFRF.time.sleep'):  # Don't actually sleep in tests
             result = getnc('test/data.ncml', 'getObs')
 
         # Should have tried 3 times
@@ -109,8 +109,8 @@ class TestGetnc:
 class TestGetObsWaveDataIntegration:
     """Integration tests for getObs.getWaveData with mocked network."""
 
-    @patch('getdatatestbed.getDataFRF.socket.gethostbyname')
-    @patch('getdatatestbed.getDataFRF.nc.Dataset')
+    @patch('murgtools.getdata.getDataFRF.socket.gethostbyname')
+    @patch('murgtools.getdata.getDataFRF.nc.Dataset')
     def test_getwavedata_returns_expected_keys(self, mock_dataset, mock_gethostbyname):
         """Test that getWaveData returns dictionary with expected keys."""
         mock_gethostbyname.return_value = '192.168.1.1'
@@ -145,10 +145,10 @@ class TestGetObsWaveDataIntegration:
         d1 = DT.datetime(2020, 1, 1, 0, 0, 0)
         d2 = DT.datetime(2020, 1, 1, 12, 0, 0)
 
-        with patch('getdatatestbed.getDataFRF.nc.num2date') as mock_num2date:
+        with patch('murgtools.getdata.getDataFRF.nc.num2date') as mock_num2date:
             mock_num2date.return_value = np.array([d1 + DT.timedelta(minutes=30*i) for i in range(24)])
 
-            from getdatatestbed.getDataFRF import getObs
+            from murgtools.getdata.getDataFRF import getObs
             obs = getObs(d1, d2)
 
             # This would require more extensive mocking to fully work,
@@ -166,10 +166,10 @@ class TestGetDataTestBedIntegration:
         d1 = DT.datetime(2020, 1, 1, 0, 0, 0)
         d2 = DT.datetime(2020, 1, 2, 0, 0, 0)
 
-        with patch('getdatatestbed.getDataFRF.nc.date2num') as mock_date2num:
+        with patch('murgtools.getdata.getDataFRF.nc.date2num') as mock_date2num:
             mock_date2num.return_value = 1577836800.0
 
-            from getdatatestbed.getDataFRF import getDataTestBed
+            from murgtools.getdata.getDataFRF import getDataTestBed
 
             tb = getDataTestBed(d1, d2)
 
@@ -183,10 +183,10 @@ class TestGetDataTestBedIntegration:
         d1 = DT.datetime(2020, 1, 1, 0, 0, 0)
         d2 = DT.datetime(2020, 1, 2, 0, 0, 0)
 
-        with patch('getdatatestbed.getDataFRF.nc.date2num') as mock_date2num:
+        with patch('murgtools.getdata.getDataFRF.nc.date2num') as mock_date2num:
             mock_date2num.return_value = 1577836800.0
 
-            from getdatatestbed.getDataFRF import getDataTestBed
+            from murgtools.getdata.getDataFRF import getDataTestBed
 
             tb = getDataTestBed(d1, d2)
